% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorcalc.R
\name{errorcalc}
\alias{errorcalc}
\title{Tabulate error}
\usage{
errorcalc(s, m, alpha = 0.05)
}
\arguments{
\item{s}{list.}

\item{m}{list.}

\item{alpha}{numeric.}
}
\value{
data.frame.
}
\description{
Tabulate error
}
