% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow}
\alias{flow}
\title{Create stream flow time series}
\usage{
flow(gage = "07261500")
}
\arguments{
\item{gage}{character. Gage ID number.}
}
\value{
A data frame.
}
\description{
Create stream flow time series from historical stream gage data.
}
