% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipassdat.R
\name{multipassdat}
\alias{multipassdat}
\title{Format sampling data for multipass package}
\usage{
multipassdat(sampdat)
}
\arguments{
\item{sampdat}{list. Simulated sampling data (see ?eflowsim::samp).}
}
\value{
list.
}
\description{
Format sampling data for multipass package
}
