% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploteffects.R
\name{ploteffects}
\alias{ploteffects}
\title{Plot flow effects}
\usage{
ploteffects(s)
}
\arguments{
\item{s}{list. Simulated population. (see ?eflowsim::sim)}
}
\value{
plot.
}
\description{
Plot flow effects on demographic rates
}
