% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotflow.R
\name{plotflow}
\alias{plotflow}
\title{Plot hydrology time-series}
\usage{
plotflow(s, simid = NA, type = "summary")
}
\arguments{
\item{s}{list. Simulated population. (see ?eflowsim::sim)}

\item{simid}{numeric vector.}

\item{type}{character.}
}
\value{
plot.
}
\description{
Plot time-series of summary statistics for stream hydrology.
}
