% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsim.R
\name{plotsim}
\alias{plotsim}
\title{Plot population time-series}
\usage{
plotsim(s, simid = NA, type = "summary")
}
\arguments{
\item{s}{list. Simulated population. (see ?eflowsim::sim)}

\item{simid}{numeric vector.}

\item{type}{character.}
}
\value{
plot.
}
\description{
Plot time-series of population and demographic rates.
}
