% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.R
\name{samp}
\alias{samp}
\title{Population samples}
\usage{
samp(
  s.dat,
  nsites = NULL,
  sitelength.min = 100,
  sitelength.max = 100,
  npass.min = 3,
  npass.max = 3,
  sigmaprop = 1,
  sigmap = 1,
  b0p = 0,
  b1p = 0,
  b2p = 0,
  b3p = 0,
  b4p = 0,
  delta = 0
)
}
\arguments{
\item{s.dat}{list. Simulated population. (see ?eflowsim::sim)}

\item{nsites}{numeric. Number of sample sites.}

\item{sitelength.min}{numeric. Minimum length of a sample site.}

\item{sitelength.max}{numeric. Maximum length of a sample site.}

\item{npass.min}{numeric. Minimum number of removal sampling passes.}

\item{npass.max}{numeric. Maximum number of removal sampling passes.}

\item{sigmaprop}{numeric. Standard deviation in proportion of total population extent sampled.}

\item{sigmap}{numeric. Standard deviation in first-pass detection probabilities.}

\item{b0p}{numeric. Regression intercept for first-pass detection probabilities.}

\item{b1p}{numeric. Effect of covariate 1 on first-pass detection probabilities.}

\item{b2p}{numeric. Effect of covariate 2 on first-pass detection probabilities.}

\item{b3p}{numeric. Effect of covariate 3 on first-pass detection probabilities.}

\item{b4p}{numeric. Effect of covariate 4 on first-pass detection probabilities.}

\item{delta}{numeric. Exponential rate of decline in sampling probabilities in each subsequent pass.}
}
\value{
list.
}
\description{
Sample from an eflowsim simulated population time series
}
