% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{eFlow population simulation}
\usage{
sim(
  covs,
  logcovs = T,
  nyears = 25,
  nsim = 1,
  densityType = "k",
  Ninit = NA,
  growth.max = Inf,
  pop.max = 1e+06,
  extent = 14,
  sigmaR = 0.01,
  b0k = 10000,
  b1k = 0,
  b2k = 1000,
  b3k = 0,
  b4k = 0,
  b0phi = -10000,
  b1phi = 0,
  b2phi = 0,
  b3phi = 0,
  b4phi = 0,
  b0r = 1,
  b1r = 0,
  b2r = 0,
  b3r = 0.5,
  b4r = 0,
  rlag = c(0, 0, 0, 0),
  dlag = c(0, 0, 0, 0)
)
}
\arguments{
\item{covs}{data.frame. Stream-based covariates (see ?eflowsim::flow).}

\item{logcovs}{logical. Log-transform covariates?}

\item{nyears}{numeric. Number of years in population time-series.}

\item{nsim}{numeric. Number of population simulations.}

\item{densityType}{character. Define density-dependence using k or phi?}

\item{Ninit}{numeric. Initial population size.}

\item{growth.max}{numeric. Maximum intrinsic population growth rate (r).}

\item{pop.max}{numeric. Maximum population size.}

\item{extent}{numeric. Spatial extent of stream network (km).}

\item{sigmaR}{numeric. Standard deviation in stochastic realized growth rates.}

\item{b0k}{numeric. Intercept for regression of carrying capacity.}

\item{b1k}{numeric. Effect of covariate 1 on carrying capacity.}

\item{b2k}{numeric. Effect of covariate 2 on carrying capacity.}

\item{b3k}{numeric. Effect of covariate 3 on carrying capacity.}

\item{b4k}{numeric. Effect of covariate 4 on carrying capacity.}

\item{b0phi}{numeric.  Intercept for regression of phi.}

\item{b1phi}{numeric.  Effect of covariate 1 on phi.}

\item{b2phi}{numeric.  Effect of covariate 1 on phi.}

\item{b3phi}{numeric.  Effect of covariate 1 on phi.}

\item{b4phi}{numeric.  Effect of covariate 1 on phi.}

\item{b0r}{numeric. Intercept for regression of intrinsic growth rate.}

\item{b1r}{numeric. Effect of covariate 1 on intrinsic growth rates.}

\item{b2r}{numeric. Effect of covariate 2 on intrinsic growth rates.}

\item{b3r}{numeric. Effect of covariate 3 on intrinsic growth rates.}

\item{b4r}{numeric. Effect of covariate 4 on intrinsic growth rates.}

\item{rlag}{numeric vector (length = 4). Time lag in years for each covariate on intrinsic growth rates.}

\item{dlag}{numeric vector (length = 4). Time lag in years for each covariate on density-dependence.}
}
\value{
list.
}
\description{
Generate an eflow population simulation
}
