\name{multipassdat}
\alias{multipassdat}
\title{Format simulated sampling data for package \pkg{multipass}}

\usage{
multipassdat()
}

\description{
Format simulated sampling data for package \pkg{multipass}
}

\arguments{

\item{argument}{ Argument description}
}

\value{
An object
}

\examples{
# Example code
}
