\name{sim}
\alias{sim}
\title{Create a simulated population time series}

\usage{
sim()
}

\description{
Create a simulated population time series with user-defined vital rates and flow effects.
}

\arguments{

\item{argument}{Argument description}
}

\value{
An object
}

\examples{
# Example code
}
